+++ src/main/java/roomescape/dto/ReservationTimeRequest.java
@@ -1,7 +1,8 @@
-package roomescape;
+package roomescape.dto;
 
 import jakarta.validation.constraints.NotNull;
 import java.time.LocalTime;
+import roomescape.domain.ReservationTime;
 
 public record ReservationTimeRequest(
         @NotNull LocalTime startAt
