+++ src/test/java/roomescape/dao/DatabaseTest.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.dao;
 
 import static org.assertj.core.api.Assertions.assertThat;
